+++ NEWS.md
@@ -1,6 +1,6 @@
 ## Package Updates
 
-### Changes in Version 1.1.0 (TBD)
+### Changes in Version 1.1.0 (2023-09-16)
 
 -   Added the skewness test of Shi et al. (2023) to the
     `describe_studies()` function